utilizzo cache + slider (da fixare)

+++ FinalAdvancedWeb/clientreact/src/index.js
@@ -7,11 +7,22 @@ class Page extends React.Component {
 
     constructor() {
         super();
-        localStorage.clear();
         this.state = {
             page: "home",
             lingua: "it",
         };
+        let data = new Date();
+        var oldData = localStorage.getItem('data');
+        if (oldData != null) {
+            oldData = new Date(oldData);
+            if (oldData.getFullYear() !== data.getFullYear() || oldData.getMonth() !== data.getMonth() || oldData.getDate() !== data.getDate()) {
+                localStorage.clear();
+                localStorage.setItem('data', data);
+            }
+        } else {
+            localStorage.clear();
+            localStorage.setItem('data', data);
+        }
         this.changePage = this.changePage.bind(this);
         this.changeLingua = this.changeLingua.bind(this);
     }
@@ -58,7 +69,7 @@ class Page extends React.Component {
             <div>
                 <Header onClick = {this.changePage}/>
                 <Nav lingua = {this.state.lingua} onPageChange = {this.changePage} onLinguaChange = {this.changeLingua}/>
-                <Home/>
+                <Home lingua = {this.state.lingua} onPageChange = {this.changePage}/>
                 <Footer lingua = {this.state.lingua} onPageChange = {this.changePage}/>
             </div>
         );
